refactor(core): drop CoordinatorConnectionModule socketConnection
Three surgical edits that finish unwiring the legacy coordinator socket
from the live code path:

1. CoordinatorConnectionModule narrows its ConnectionModuleDeclaration
   generic to Unit and turns `socketConnection` into a `Unit = Unit`
   vestigial override. The CoordinatorSocketConnection field is deleted;
   the class file itself stays on disk until the legacy socket interfaces
   are removed later in v2.

2. StreamLifecycleObserver.kt moved from `lifecycle/` to `lifecycle/legacy/`.
   Coordinator-side lifecycle observation now flows through core's
   StreamLifecycleMonitor. Legacy consumers (Coordinator/SFU socket files
   kept for their own later migration) have their imports updated to
   `lifecycle.legacy.StreamLifecycleObserver`. The SFU migrates separately;
   the moved file survives until then.

3. IntegrationTestBase updated to read the connection ID from
   streamClient.connectionState instead of the deleted socketConnection.

Legacy coordinator test @Ignore roster:
- CoordinatorSocketTest.kt / CoordinatorSocketStateServiceTest.kt do NOT
  exist under `src/test/.../socket/coordinator/` on this branch — nothing
  to annotate.
- TokenManagerImplTest.kt exercises TokenManagerImpl behaviours shared with
  the SFU (still in-tree). Preserved unannotated so SFU coverage is not
  lost.

@Ignore baseline unchanged at 21.
